完善实体类，优化导入逻辑

+++ src/main/kotlin/com/zhile/excelutil/dao/ImDepartmentRMSRepository.kt
@@ -3,4 +3,4 @@ package com.zhile.excelutil.dao
 import com.zhile.excelutil.entity.ImDepartmentRMS
 import org.springframework.data.repository.CrudRepository
 
-interface ImDepartmentRMSRepository : CrudRepository<ImDepartmentRMS, Int>
+interface ImDepartmentRMSRepository : CrudRepository<ImDepartmentRMS, Long>
