+++ src/main/kotlin/com/zhile/excelutil/dao/ImDepartmentRepository.kt
@@ -3,4 +3,4 @@ package com.zhile.excelutil.dao
 import com.zhile.excelutil.entity.ImDepartment
 import org.springframework.data.repository.CrudRepository
 
-interface ImDepartmentRepository : CrudRepository<ImDepartment, Int>
+interface ImDepartmentRepository : CrudRepository<ImDepartment, Long>
